Added options for more perfect filling (Issue #1)
We now deconstruct the leftover cores so that they are distributed
perfectly evenly, possibly over multiple clusters.

Say we have a problem where we want something with a remainder of 16 on
a cluster with 24 cores per node. It will now distribute these 16 cores
so that for every 3 runs, they will each have 8 cores on two separate
leftover nodes.

One could also try to place them more "optimally", with the least amount
of spread. This is however a hard problem to solve, and will not
necessarily give an even runtime as some jobs will have more spread than
others. This approach is both easy and "fair", however much that is
worth.

Closes #1